sweekruth.somaraju/debitcards/debitcards/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
azure-eventhub==5.11.5
python-jose==3.3.0
python-multipart==0.0.6
